refactor download logic
