+++ shared/src/commonMain/kotlin/com/dnfapps/arrmatey/bazarr/viewmodel/BazarrDetailsViewModel.kt
@@ -5,6 +5,7 @@ import androidx.lifecycle.viewModelScope
 import com.dnfapps.arrmatey.bazarr.api.model.BazarrMediaType
 import com.dnfapps.arrmatey.bazarr.api.model.BazarrSubtitle
 import com.dnfapps.arrmatey.bazarr.state.BazarrDetails
+import com.dnfapps.arrmatey.bazarr.usecase.DownloadBazarrSubtitleToDeviceUseCase
 import com.dnfapps.arrmatey.bazarr.usecase.GetBazarrEpisodesUseCase
 import com.dnfapps.arrmatey.bazarr.usecase.GetBazarrMediaDetailsUseCase
 import com.dnfapps.arrmatey.bazarr.usecase.PerformBazarrAutomaticSearchUseCase
@@ -12,7 +13,6 @@ import com.dnfapps.arrmatey.client.OperationStatus
 import com.dnfapps.arrmatey.client.onError
 import com.dnfapps.arrmatey.client.onSuccess
 import com.dnfapps.arrmatey.instances.usecase.GetBazarrInstanceRepositoryUseCase
-import com.dnfapps.arrmatey.notifications.NotificationManager
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
@@ -29,7 +29,7 @@ class BazarrDetailsViewModel(
     private val getBazarrEpisodesUseCase: GetBazarrEpisodesUseCase,
     private val getBazarrInstanceRepositoryUseCase: GetBazarrInstanceRepositoryUseCase,
     private val performBazarrAutomaticSearchUseCase: PerformBazarrAutomaticSearchUseCase,
-    private val notificationManager: NotificationManager
+    private val downloadBazarrSubtitleToDeviceUseCase: DownloadBazarrSubtitleToDeviceUseCase
 ) : ViewModel() {
 
     private val _operationState = MutableStateFlow<OperationStatus>(OperationStatus.Idle)
@@ -64,35 +64,8 @@ class BazarrDetailsViewModel(
     }
 
     fun downloadToDevice(subtitle: BazarrSubtitle, onResult: (ByteArray?) -> Unit) {
-        val path = subtitle.path ?: return
-        val fileName = path.substringAfterLast('/')
         viewModelScope.launch {
-            val repo = getBazarrInstanceRepositoryUseCase.observeSelected().firstOrNull() ?: return@launch
-            val notificationId = path.hashCode()
-            val instanceName = repo.instance.label
-
-            repo.getSubtitleFile(path) { progress ->
-                notificationManager.showProgressNotification(
-                    id = notificationId,
-                    title = "Downloading $fileName",
-                    message = "In progress...",
-                    progress = progress,
-                    instanceName = instanceName
-                )
-            }
-                .onSuccess {
-                    notificationManager.showNotification(
-                        id = notificationId,
-                        title = "Download complete",
-                        message = fileName,
-                        instanceName = instanceName
-                    )
-                    onResult(it)
-                }
-                .onError { _, _, _ ->
-                    notificationManager.cancelNotification(notificationId)
-                    onResult(null)
-                }
+            downloadBazarrSubtitleToDeviceUseCase(subtitle, onResult)
         }
     }
 }
