+++ shared/src/commonMain/kotlin/com/dnfapps/arrmatey/bazarr/viewmodel/BazarrMediaSubtitlesViewModel.kt
@@ -6,13 +6,11 @@ import com.dnfapps.arrmatey.bazarr.api.model.BazarrSubtitle
 import com.dnfapps.arrmatey.bazarr.api.model.BazarrSubtitleLanguage
 import com.dnfapps.arrmatey.bazarr.state.BazarrMediaTarget
 import com.dnfapps.arrmatey.bazarr.state.BazarrSubtitlesUiState
+import com.dnfapps.arrmatey.bazarr.usecase.DownloadBazarrSubtitleToDeviceUseCase
 import com.dnfapps.arrmatey.client.OperationStatus
 import com.dnfapps.arrmatey.client.onError
 import com.dnfapps.arrmatey.client.onSuccess
 import com.dnfapps.arrmatey.instances.usecase.GetBazarrInstanceRepositoryUseCase
-import com.dnfapps.arrmatey.notifications.NotificationManager
-import com.dnfapps.arrmatey.shared.MR
-import com.dnfapps.arrmatey.utils.MokoStrings
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
@@ -27,8 +25,7 @@ import kotlinx.coroutines.launch
 class BazarrMediaSubtitlesViewModel(
     private val target: BazarrMediaTarget,
     private val getBazarrInstanceRepositoryUseCase: GetBazarrInstanceRepositoryUseCase,
-    private val notificationManager: NotificationManager,
-    private val mokoStrings: MokoStrings
+    private val downloadBazarrSubtitleToDeviceUseCase: DownloadBazarrSubtitleToDeviceUseCase
 ) : ViewModel() {
 
     private val _state = MutableStateFlow<BazarrSubtitlesUiState>(BazarrSubtitlesUiState.Loading)
@@ -136,35 +133,8 @@ class BazarrMediaSubtitlesViewModel(
     }
 
     fun downloadToDevice(subtitle: BazarrSubtitle, onResult: (ByteArray?) -> Unit) {
-        val path = subtitle.path ?: return
-        val fileName = path.substringAfterLast('/')
         viewModelScope.launch {
-            val repo = repo() ?: return@launch
-            val notificationId = path.hashCode()
-            val instanceName = repo.instance.label
-
-            repo.getSubtitleFile(path) { progress ->
-                notificationManager.showProgressNotification(
-                    id = notificationId,
-                    title = mokoStrings.getString(MR.strings.downloading_file, listOf(fileName)),
-                    message = mokoStrings.getString(MR.strings.downloading_progress),
-                    progress = progress,
-                    instanceName = instanceName
-                )
-            }
-                .onSuccess {
-                    notificationManager.showNotification(
-                        id = notificationId,
-                        title = mokoStrings.getString(MR.strings.download_complete),
-                        message = fileName,
-                        instanceName = instanceName
-                    )
-                    onResult(it)
-                }
-                .onError { _, _, _ ->
-                    notificationManager.cancelNotification(notificationId)
-                    onResult(null)
-                }
+            downloadBazarrSubtitleToDeviceUseCase(subtitle, onResult)
         }
     }
